{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-q2>=1.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=24.1.1",
    "psycopg2-binary>=2.9.11",
    "pyotp>=2.9.0",
    "uvicorn>=0.40.0",
]
